Update the End User License Agreement to more clearly state the intention of the limited commercial use.
The License limitation clause permitted the license for only personal and non-commercial use, however, I pointed out that any streamer or video produced using this application for a channel that receives monetary compensation via ads or subs or bits or superchats or donations was, in fact, actually commercial use that was prohibited by this license.  I was talking with cyanidesugar, and, they communicated that their intention with the license was not to prevent a streamer or video maker from using the panel on their channel but to prevent people from taking the software and incorporating it into a cloud service commercially. I have modified the license with my suggestions that carve out exactly that use case in the License Limitation section and clarifies the conditions where the software is allowed to be used in a small but important business/commercial use case.  This application is no longer being developed, so I'm not expecting this to be merged, it's simply a suggestion if ever this license was going to be re-used by a developer in the future.